requirements.txt
django==2.0.7
djangorestframework